Fix update

+++ src/main/java/org/example/services/CategoryService.java
@@ -1,7 +1,9 @@
 package org.example.services;
 
 import org.example.dtos.CategoryDto;
+import org.example.dtos.InvoiceDto;
 import org.example.entities.Category;
+import org.example.entities.Invoice;
 import org.example.exceptions.CategoryException;
 import org.example.interfaces.ICategoryRepository;
 import org.example.interfaces.ICategoryService;
@@ -10,8 +12,10 @@ import org.example.mapping.CategoryMapper;
 import org.example.models.FileFormats;
 import org.example.models.CategoryCreationModel;
 import org.example.models.CategoryResponse;
+import org.example.models.InvoiceResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Page;
 
@@ -50,6 +54,14 @@ public class CategoryService implements ICategoryService {
         return new CategoryResponse(mapper.toDto(categoryPage.getContent()),categoryPage.getTotalElements());
     }
 
+    public CategoryResponse getCategories(int page,int size) {
+        PageRequest pageRequest = PageRequest.of(
+                page, size);
+        Page<Category> categoriesPage = repo.findAll(pageRequest);
+        Iterable<CategoryDto> categories = mapper.toDto(categoriesPage.getContent());
+        return  new CategoryResponse(categories,categoriesPage.getTotalElements());
+    }
+
     @Override
     public CategoryDto getCategoryById(Long id) {
         Optional<Category> category = repo.findById(id);
@@ -79,8 +91,9 @@ public class CategoryService implements ICategoryService {
         boolean isPresent = optCategory.isPresent();
         if(isPresent){
             Category category = mapper.fromCreationModel(categoryModel);
+            category.setImage(optCategory.get().getImage());
             category.setCreationTime(LocalDateTime.now());
-            if(!categoryModel.getFile().isEmpty() ){
+            if(categoryModel.getFile()!=null && !categoryModel.getFile().isEmpty() ){
                 storageService.deleteImage(optCategory.get().getImage());
                 String imageName = storageService.saveImage(categoryModel.getFile(),FileFormats.JPG);
                 category.setImage(imageName);
